SkinAtlas: restore (uint) cast around Mathf.Log
Option C (move SkinPreview to layer 10, default 1<<10) only fixes pristine inspector state. Existing serialized m_Layer values can still be problematic: the SkinAtlas.prefab has m_Layer: 31 left over from when the field was `int` (storing layer index), which as a LayerMask reads as bits 0-4 set; scene overrides may have other historical values.

Casting through uint before Mathf.Log handles every case: legacy int values, sign-bit-31 negatives, and multi-bit masks. Costs three lines and a comment; eliminates a whole class of "layer out of range" errors.

+++ Draw.ioTest/Draw.ioTest/Assets/Scripts/UI/SkinAtlas.cs
@@ -40,7 +40,10 @@ public class SkinAtlas : MonoBehaviour
         m_Camera.orthographicSize = Rows * m_CellWorldSize * 0.5f;
         m_Camera.aspect           = (float)m_Columns / Rows;
 
-        int layer = (int)Mathf.Log(m_Layer.value, 2);
+        // (uint) cast handles legacy int values stored before the field type
+        // change to LayerMask, plus the sign-bit-31 edge case where
+        // m_Layer.value reads negative.
+        int layer = (int)Mathf.Log((uint)m_Layer.value, 2);
 
         for (int i = 0; i < count; i++)
         {
